a global variable reference is an expression result

+++ test/c1.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 
-short v = 123;
+int v = 123;
 
 int main(int argc, char** args) {
     printf( "%d\n", v );
